removed ip parsing
apparently this TcpListener instance line is obsolete, it now requires you to parse the users IP and send it along with the port as an argument to the TcpListener function. Still works, and for the purposes of this project I will leave it as it is. Any solutions or suggestions for the sake of knowledge and future implementations would be great.